requirements.txt
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14
google-genai==0.3.0
python-dotenv==1.0.1
Flask==3.0.2
gunicorn==21.2.0
PyPDF2==3.0.1
Werkzeug==3.0.1
